dashboard_control: carve api_transfers_fs.rs (pure move)
Transfer jobs and the scoped filesystem surface. The shared
test_upload_state fixture stays pinned in mod.rs tests (api_media
already imports it from there; the closure rule is blind to
cross-slice tests-imports and tried to move it with its last
same-file referrers).
